refactor(editor): extract useCropTool (Phase 3a)
Move the crop/transform tool out of EditorOverlay — the biggest tangle: pointer
interactions (resize/rotate/pan handles, straighten-ruler drag), the commit ops
(aspect ratio, quarter-turn/flip, free angle), the wheel-zoom wiring, and the
re-clamp-on-geometry-change effect. useCropTool owns activeInteraction + the
in-flight pointer/angle drag refs, and takes the viewport geometry + history it
operates on.

Supporting moves:
- createCenteredCrop / symmetricResize + MIN_FREE_ANGLE / MAX_FREE_ANGLE → cropMath.
- useEditorHistory gains commitCurrent() (pushes the live state — the drag-end
  history commit), so the crop hook no longer touches history internals.
- Dropped stray debug console.log/warn from the moved commit paths.

EditorOverlay: 1498 → 1218 lines. Behavior-preserving: 18-editor-transform
(rotate/flip/aspect + undo/redo) + 03/06/19/20 green; full suite clean aside from
the two known contention flakes (pass in isolation).
